--- AuthAPI/Contracts/Requests/RegisterRequest.cs
@@ -1,11 +1,0 @@
-﻿using AuthAPI.Models;
-
-namespace AuthAPI.Contracts.Requests
-{
-    public class RegisterRequest
-    {
-        public string Email { get; set; }
-        public string Password { get; set; }
-        public Role Role { get; set; }
-    }
-}
