+++ MyPackage/JuliaChistyakovaPackage.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    <PackageTags>Ports; Microservices; Contracts.Response; Contracts.Request</PackageTags>
+    <Title>JuliaChistyakovaPackage</Title>
+    <PackageId>JuliaChistyakovaPackage</PackageId>
+    <Version>1.0.1</Version>
+    <Authors>klepaski</Authors>
+  </PropertyGroup>
+
+</Project>
