+++ Orchestrator/Controllers/OrchestratorController.cs
@@ -1,14 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
-using Orchestrator.Contracts.Requests;
 using Orchestrator.Services;
-using System;
-using System.Net.Http;
-using System.Threading.Tasks;
-using System.Transactions;
-using Newtonsoft.Json;
-using Orchestrator.Contracts.Responses;
-using Orchestrator;
-using System.Text;
+using JuliaChistyakovaPackage;
 
 namespace Orchestrator.Controllers
 {
