--- Orchestrator/Ports.cs
@@ -1,8 +1,0 @@
-﻿namespace Orchestrator
-{
-    public static class Ports
-    {
-        public const string AuthAPI = "http://localhost:5001";
-        public const string ProfilesAPI = "http://localhost:5003";
-    }
-}
