+++ ProfilesAPI/Controllers/AccountController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using ProfilesAPI.Contracts.Requests;
 using ProfilesAPI.Services;
+using JuliaChistyakovaPackage;
 
 namespace ProfilesAPI.Controllers
 {
